+++ node/src/main/kotlin/net/corda/node/services/config/NodeConfiguration.kt
@@ -25,7 +25,7 @@ interface NodeConfiguration : NodeSSLConfiguration {
     val rpcUsers: List<User>
     val devMode: Boolean
     val devModeOptions: DevModeOptions?
-    val certificateSigningService: URL
+    val compatibilityZoneURL: URL?
     val certificateChainCheckPolicies: List<CertChainPolicyConfig>
     val verifierType: VerifierType
     val messageRedeliveryDelaySeconds: Int
@@ -89,7 +89,7 @@ data class NodeConfigurationImpl(
         override val trustStorePassword: String,
         override val dataSourceProperties: Properties,
         override val database: Properties?,
-        override val certificateSigningService: URL,
+        override val compatibilityZoneURL: URL? = null,
         override val rpcUsers: List<User>,
         override val verifierType: VerifierType,
         // TODO typesafe config supports the notion of durations. Make use of that by mapping it to java.time.Duration.